Lots of cleanup with Settings

+++ app/src/main/java/me/ccrama/redditslide/Activities/Settings.java
@@ -348,7 +348,7 @@ public class Settings extends BaseActivity {
             });
         } else {
             findViewById(R.id.reddit_settings).setEnabled(false);
-            findViewById(R.id.reddit_settings).setAlpha((float) 0.25);
+            findViewById(R.id.reddit_settings).setAlpha(0.25f);
         }
 
     }
