+++ app/src/main/java/me/ccrama/redditslide/Activities/SettingsGeneral.java
@@ -265,7 +265,7 @@ public class SettingsGeneral extends BaseActivityAnim implements FolderChooserDi
             });
         } else {
             findViewById(R.id.notifications).setEnabled(false);
-            findViewById(R.id.notifications).setAlpha((float) 0.25);
+            findViewById(R.id.notifications).setAlpha(0.25f);
         }
 
         ((TextView) findViewById(R.id.sorting_current)).setText(Reddit.getSortingStrings(getBaseContext())[Reddit.getSortingId("")]);
